+++ randovania/games/am2r/gui/preset_settings/am2r_room_design_tab.py
@@ -15,18 +15,18 @@ if typing.TYPE_CHECKING:
     from randovania.interface_common.preset_editor import PresetEditor
     from randovania.layout.preset import Preset
 
-_FIELDS = [
-    "septogg_helpers",
-    "respawn_bomb_blocks",
-    "grave_grotto_blocks",
-    "nest_pipes",
-    "softlock_prevention_blocks",
-    "a3_entrance_blocks",
-    "screw_blocks",
-]
-
 
 class PresetAM2RRoomDesign(PresetTab, Ui_PresetAM2RRoomDesign):
+    _CHECKBOX_FIELDS = [
+        "septogg_helpers",
+        "respawn_bomb_blocks",
+        "grave_grotto_blocks",
+        "nest_pipes",
+        "softlock_prevention_blocks",
+        "a3_entrance_blocks",
+        "screw_blocks",
+    ]
+
     def __init__(self, editor: PresetEditor, game_description: GameDescription, window_manager: WindowManager):
         super().__init__(editor, game_description, window_manager)
         self.setupUi(self)
@@ -34,7 +34,7 @@ class PresetAM2RRoomDesign(PresetTab, Ui_PresetAM2RRoomDesign):
         self.setCentralWidget(self.root_widget)
 
         # Signals
-        for f in _FIELDS:
+        for f in self._CHECKBOX_FIELDS:
             self._add_persist_option(getattr(self, f"{f}_check"), f)
 
     @classmethod
@@ -55,5 +55,5 @@ class PresetAM2RRoomDesign(PresetTab, Ui_PresetAM2RRoomDesign):
     def on_preset_changed(self, preset: Preset):
         config = preset.configuration
         assert isinstance(config, AM2RConfiguration)
-        for f in _FIELDS:
+        for f in self._CHECKBOX_FIELDS:
             typing.cast(QtWidgets.QCheckBox, getattr(self, f"{f}_check")).setChecked(getattr(config, f))
